Refactor theme architecture and improve code quality
- Merge theme/theme.ts into lib/theme.ts to eliminate split-brain theming
- Refactor types/mui.d.ts to reduce redundancy with CustomColors interface
- Fix contrast accessibility in WorkoutSummary.tsx status chip
- Optimize HeartRateTimeSeries.tsx by memoizing tick formatter and removing unused prop
- Update all references and docs to use lib/theme.ts
