+++ app/client/experimental/components/HeartRateTimeSeries.tsx
@@ -1,4 +1,3 @@
-// app/client/experimental/components/HeartRateTimeSeries.tsx
 'use client'
 import { Card, CardContent, Typography, Box } from '@mui/material'
 import { HrDataPoint } from '@/lib/workout-session-storage'
@@ -18,6 +17,14 @@ interface HeartRateTimeSeriesProps {
   hrHistory: HrDataPoint[]
 }
 
+const timeFormatter = new Intl.DateTimeFormat('en-US', {
+  hour: 'numeric',
+  minute: 'numeric',
+  second: 'numeric',
+})
+
+const formatTime = (time: number) => timeFormatter.format(time)
+
 const HeartRateTimeSeries = ({ hrHistory }: HeartRateTimeSeriesProps) => {
   const theme = useTheme()
 
@@ -29,15 +36,12 @@ const HeartRateTimeSeries = ({ hrHistory }: HeartRateTimeSeriesProps) => {
         </Typography>
         <Box sx={{ height: 300 }} data-testid="hr-time-series-chart">
           <ResponsiveContainer width="100%" height="100%">
-            <LineChart data={hrHistory} syncId="anyId">
+            <LineChart data={hrHistory}>
               <CartesianGrid
                 stroke={theme.palette.divider}
                 strokeDasharray="3 3"
               />
-              <XAxis
-                dataKey="time"
-                tickFormatter={(time) => new Date(time).toLocaleTimeString()}
-              />
+              <XAxis dataKey="time" tickFormatter={formatTime} />
               <YAxis domain={['auto', 'auto']} />
               <Tooltip />
               <Legend />
