+++ app/client/experimental/page.tsx
@@ -3,7 +3,7 @@ import ExperimentalAnalyticsPage from './components/ExperimentalAnalyticsPage'
 import { WebSocketProvider } from '@/context/WebSocketContext'
 import { UserSettingsProvider } from '@/context/UserSettingsContext'
 import { CssBaseline, ThemeProvider } from '@mui/material'
-import theme from '@/theme/theme'
+import theme from '@/lib/theme'
 
 export default function ExperimentalPage() {
   return (
