+++ lib/theme.ts
@@ -82,10 +82,21 @@ const theme = createTheme({
     },
     // Dividers
     divider: '#E0E0E0',
+
+    // Custom status colors
+    custom: {
+      prepare: '#f59e0b',
+      work: '#ef4444',
+      rest: '#22c55e',
+      running: '#3b82f6',
+      idle: '#6b7280',
+      cooldown: '#6b7280',
+    },
   },
 
   // Typography - Clear hierarchy
   typography: {
+    fontFamilyMono: '"Roboto Mono", "Courier New", monospace',
     fontFamily: [
       '-apple-system',
       'BlinkMacSystemFont',
